extracted Visibility behaviour to a separate class
