--- unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NavMap/NavMapScrollRect.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class NavMapScrollRect : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start() { }
-
-    // Update is called once per frame
-    void Update() { }
-}
