+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NavMap/NavmapView.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 41c277d3ff1164a5fb894b76c718d96a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 92eadf420ad9498593fa1f1a17f8de19
+timeCreated: 1665078515
