+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NavMap/NavmapVisibility.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6565bd7a51f47f74eac7afd84299f58f
+guid: 41c277d3ff1164a5fb894b76c718d96a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
